Added related films and tabs
