+++ markup/add-review.html
@@ -5,7 +5,7 @@
     <title>WTW</title>
     <meta name="robots" content="noindex, nofollow">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="../project/public/css/main.min.css">
+    <link rel="stylesheet" href="css/main.min.css">
   </head>
 
   <body>
@@ -45,7 +45,7 @@
     <section class="film-card film-card--full">
       <div class="film-card__header">
          <div class="film-card__bg">
-          <img src="../project/public/img/bg-the-grand-budapest-hotel.jpg" alt="The Grand Budapest Hotel" />
+          <img src="img/bg-the-grand-budapest-hotel.jpg" alt="The Grand Budapest Hotel" />
         </div>
 
         <h1 class="visually-hidden">WTW</h1>
@@ -73,7 +73,7 @@
           <ul class="user-block">
             <li class="user-block__item">
               <div class="user-block__avatar">
-                <img src="../project/public/img/avatar.jpg" alt="User avatar" width="63" height="63" />
+                <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
               </div>
             </li>
             <li class="user-block__item">
@@ -83,7 +83,7 @@
         </header>
 
         <div class="film-card__poster film-card__poster--small">
-          <img src="../project/public/img/the-grand-budapest-hotel-poster.jpg" alt="The Grand Budapest Hotel poster" width="218" height="327" />
+          <img src="img/the-grand-budapest-hotel-poster.jpg" alt="The Grand Budapest Hotel poster" width="218" height="327" />
         </div>
       </div>
