+++ markup/head-guest.html
@@ -5,7 +5,7 @@
     <title>WTW</title>
     <meta name="robots" content="noindex, nofollow">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="../project/public/css/main.min.css">
+    <link rel="stylesheet" href="css/main.min.css">
   </head>
 
   <body>
@@ -45,7 +45,7 @@
     <section class="film-card">
       <div class="film-card__bg">
         <!-- <img src="img/bg-the-grand-budapest-hotel.jpg" alt="The Grand Budapest Hotel" /> -->
-        <img src="../project/public/img/bg-header.jpg" />
+        <img src="img/bg-header.jpg" />
       </div>
 
       <h1 class="visually-hidden">WTW</h1>
