+++ markup/movie-page.html
@@ -5,7 +5,7 @@
     <title>WTW</title>
     <meta name="robots" content="noindex, nofollow">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="../project/public/css/main.min.css">
+    <link rel="stylesheet" href="css/main.min.css">
   </head>
 
   <body>
@@ -48,7 +48,7 @@
     <section class="film-card film-card--full">
       <div class="film-card__hero">
         <div class="film-card__bg">
-          <img src="../project/public/img/bg-the-grand-budapest-hotel.jpg" alt="The Grand Budapest Hotel" />
+          <img src="img/bg-the-grand-budapest-hotel.jpg" alt="The Grand Budapest Hotel" />
         </div>
 
         <h1 class="visually-hidden">WTW</h1>
@@ -65,7 +65,7 @@
           <ul class="user-block">
             <li class="user-block__item">
               <div class="user-block__avatar">
-                <img src="../project/public/img/avatar.jpg" alt="User avatar" width="63" height="63" />
+                <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
               </div>
             </li>
             <li class="user-block__item">
@@ -105,7 +105,7 @@
       <div class="film-card__wrap film-card__translate-top">
         <div class="film-card__info">
           <div class="film-card__poster film-card__poster--big">
-            <img src="../project/public/img/the-grand-budapest-hotel-poster.jpg" alt="The Grand Budapest Hotel poster" width="218" height="327" />
+            <img src="img/the-grand-budapest-hotel-poster.jpg" alt="The Grand Budapest Hotel poster" width="218" height="327" />
           </div>
 
           <div class="film-card__desc">
@@ -152,7 +152,7 @@
         <div class="catalog__films-list">
           <article class="small-film-card catalog__films-card">
             <div class="small-film-card__image">
-              <img src="../project/public/img/fantastic-beasts-the-crimes-of-grindelwald.jpg" alt="Fantastic Beasts: The Crimes of Grindelwald" width="280" height="175" />
+              <img src="img/fantastic-beasts-the-crimes-of-grindelwald.jpg" alt="Fantastic Beasts: The Crimes of Grindelwald" width="280" height="175" />
             </div>
             <h3 class="small-film-card__title">
               <a class="small-film-card__link" href="film-page.html">Fantastic Beasts: The Crimes of Grindelwald</a>
@@ -161,7 +161,7 @@
 
           <article class="small-film-card catalog__films-card">
             <div class="small-film-card__image">
-              <img src="../project/public/img/bohemian-rhapsody.jpg" alt="Bohemian Rhapsody" width="280" height="175" />
+              <img src="img/bohemian-rhapsody.jpg" alt="Bohemian Rhapsody" width="280" height="175" />
             </div>
             <h3 class="small-film-card__title">
               <a class="small-film-card__link" href="film-page.html">Bohemian Rhapsody</a>
@@ -170,7 +170,7 @@
 
           <article class="small-film-card catalog__films-card">
             <div class="small-film-card__image">
-              <img src="../project/public/img/macbeth.jpg" alt="Macbeth" width="280" height="175" />
+              <img src="img/macbeth.jpg" alt="Macbeth" width="280" height="175" />
             </div>
             <h3 class="small-film-card__title">
               <a class="small-film-card__link" href="film-page.html">Macbeth</a>
@@ -179,7 +179,7 @@
 
           <article class="small-film-card catalog__films-card">
             <div class="small-film-card__image">
-              <img src="../project/public/img/aviator.jpg" alt="Aviator" width="280" height="175" />
+              <img src="img/aviator.jpg" alt="Aviator" width="280" height="175" />
             </div>
             <h3 class="small-film-card__title">
               <a class="small-film-card__link" href="film-page.html">Aviator</a>
