+++ markup/my-list.html
@@ -5,7 +5,7 @@
     <title>WTW</title>
     <meta name="robots" content="noindex, nofollow">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="../project/public/css/main.min.css">
+    <link rel="stylesheet" href="css/main.min.css">
   </head>
 
   <body>
@@ -56,7 +56,7 @@
         <ul class="user-block">
           <li class="user-block__item">
             <div class="user-block__avatar">
-              <img src="../project/public/img/avatar.jpg" alt="User avatar" width="63" height="63" />
+              <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
             </div>
           </li>
           <li class="user-block__item">
@@ -71,7 +71,7 @@
         <div class="catalog__films-list">
           <article class="small-film-card catalog__films-card">
             <div class="small-film-card__image">
-              <img src="../project/public/img/fantastic-beasts-the-crimes-of-grindelwald.jpg" alt="Fantastic Beasts: The Crimes of Grindelwald" width="280" height="175" />
+              <img src="img/fantastic-beasts-the-crimes-of-grindelwald.jpg" alt="Fantastic Beasts: The Crimes of Grindelwald" width="280" height="175" />
             </div>
             <h3 class="small-film-card__title">
               <a class="small-film-card__link" href="film-page.html">Fantastic Beasts: The Crimes of Grindelwald</a>
@@ -80,7 +80,7 @@
 
           <article class="small-film-card catalog__films-card">
             <div class="small-film-card__image">
-              <img src="../project/public/img/bohemian-rhapsody.jpg" alt="Bohemian Rhapsody" width="280" height="175" />
+              <img src="img/bohemian-rhapsody.jpg" alt="Bohemian Rhapsody" width="280" height="175" />
             </div>
             <h3 class="small-film-card__title">
               <a class="small-film-card__link" href="film-page.html">Bohemian Rhapsody</a>
@@ -89,7 +89,7 @@
 
           <article class="small-film-card catalog__films-card">
             <div class="small-film-card__image">
-              <img src="../project/public/img/macbeth.jpg" alt="Macbeth" width="280" height="175" />
+              <img src="img/macbeth.jpg" alt="Macbeth" width="280" height="175" />
             </div>
             <h3 class="small-film-card__title">
               <a class="small-film-card__link" href="film-page.html">Macbeth</a>
@@ -98,7 +98,7 @@
 
           <article class="small-film-card catalog__films-card">
             <div class="small-film-card__image">
-              <img src="../project/public/img/aviator.jpg" alt="Aviator" width="280" height="175" />
+              <img src="img/aviator.jpg" alt="Aviator" width="280" height="175" />
             </div>
             <h3 class="small-film-card__title">
               <a class="small-film-card__link" href="film-page.html">Aviator</a>
@@ -108,7 +108,7 @@
 
           <article class="small-film-card catalog__films-card">
             <div class="small-film-card__image">
-              <img src="../project/public/img/we-need-to-talk-about-kevin.jpg" alt="We need to talk about Kevin" width="280" height="175" />
+              <img src="img/we-need-to-talk-about-kevin.jpg" alt="We need to talk about Kevin" width="280" height="175" />
             </div>
             <h3 class="small-film-card__title">
               <a class="small-film-card__link" href="film-page.html">We need to talk about Kevin</a>
@@ -117,7 +117,7 @@
 
           <article class="small-film-card catalog__films-card">
             <div class="small-film-card__image">
-              <img src="../project/public/img/what-we-do-in-the-shadows.jpg" alt="What We Do in the Shadows" width="280" height="175" />
+              <img src="img/what-we-do-in-the-shadows.jpg" alt="What We Do in the Shadows" width="280" height="175" />
             </div>
             <h3 class="small-film-card__title">
               <a class="small-film-card__link" href="film-page.html">What We Do in the Shadows</a>
@@ -126,7 +126,7 @@
 
           <article class="small-film-card catalog__films-card">
             <div class="small-film-card__image">
-              <img src="../project/public/img/revenant.jpg" alt="Revenant" width="280" height="175" />
+              <img src="img/revenant.jpg" alt="Revenant" width="280" height="175" />
             </div>
             <h3 class="small-film-card__title">
               <a class="small-film-card__link" href="film-page.html">Revenant</a>
@@ -135,7 +135,7 @@
 
           <article class="small-film-card catalog__films-card">
             <div class="small-film-card__image">
-              <img src="../project/public/img/johnny-english.jpg" alt="Johnny English" width="280" height="175" />
+              <img src="img/johnny-english.jpg" alt="Johnny English" width="280" height="175" />
             </div>
             <h3 class="small-film-card__title">
               <a class="small-film-card__link" href="film-page.html">Johnny English</a>
@@ -145,7 +145,7 @@
 
           <article class="small-film-card catalog__films-card">
             <div class="small-film-card__image">
-              <img src="../project/public/img/shutter-island.jpg" alt="Shutter Island" width="280" height="175" />
+              <img src="img/shutter-island.jpg" alt="Shutter Island" width="280" height="175" />
             </div>
             <h3 class="small-film-card__title">
               <a class="small-film-card__link" href="film-page.html">Shutter Island</a>
