+++ markup/player-pause.html
@@ -5,7 +5,7 @@
     <title>WTW</title>
     <meta name="robots" content="noindex, nofollow">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="../project/public/css/main.min.css">
+    <link rel="stylesheet" href="css/main.min.css">
   </head>
 
   <body>
@@ -43,7 +43,7 @@
     </div>
 
     <div class="player">
-      <video src="#" class="player__video" poster="../project/public/img/player-poster.jpg"></video>
+      <video src="#" class="player__video" poster="img/player-poster.jpg"></video>
 
       <button type="button" class="player__exit">Exit</button>
