+++ markup/sign-in.html
@@ -5,7 +5,7 @@
     <title>WTW</title>
     <meta name="robots" content="noindex, nofollow">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="../project/public/css/main.min.css">
+    <link rel="stylesheet" href="css/main.min.css">
   </head>
 
   <body>
