+++ project/src/components/film-card/film-card.tsx
@@ -24,11 +24,13 @@ export default function FilmCard ({film}:FilmCardProps): JSX.Element{
       }}
     >
       {(isActive && <Video src={film.prevVideoLink} isMuted posterIMG={film.posterImage}/>) ||
-      <div className="small-film-card__image">
-        <img src={film.prevImage} alt={film.name} width="280" height="175"/>
-      </div>}
-      <h3 className="small-film-card__title">
-        <Link className="small-film-card__link" to={`/films/${film.id}`}>{film.name}</Link>
-      </h3>
+      <>
+        <div className="small-film-card__image">
+          <img src={film.prevImage} alt={film.name} width="280" height="175"/>
+        </div>
+        <h3 className="small-film-card__title">
+          <Link className="small-film-card__link" to={`/films/${film.id}`}>{film.name}</Link>
+        </h3>
+      </>}
     </article>);
 }
